juksutil: Fix problem with time zone Asia/Tehran
It has transitions already after year 2038 so have to be careful with
the data types.

+++ juksutil/src/tz.c
@@ -129,7 +129,7 @@ static bool find_next_transition(tzinfo_t *dest, char *p, int len, time_t now)
 	while (low+1 != high) {
 		uint32_t mid = (high+low)/2;
 		printf("low %d mid %d high %d\n", low, mid, high);
-		int32_t ts_mid = be64toh(transition[mid]);
+		int64_t ts_mid = be64toh(transition[mid]);
 		if (now < ts_mid) {
 			high = mid;
 		} else {
